Rename AbstractMHQDialog to AbstractMHQDialogBasic
Updated references and class declarations across the codebase to reflect the new name. This improves clarity by better describing the purpose of the base class within the GUI structure.
